Huny-B.com

+++ scripts/init-subs.sh
@@ -8,14 +8,20 @@
 
 # Copyright (c) 2018 Davsk. All Rights Reserved.
 
+# identify windows operating system. set windows variable.
 windows() { [[ -n "$WINDIR" ]]; }
+
+# if not exists web/themes/.git then clone hugoThemes.git 
+if [[ ! -d "web/themes/.git" ]]; then
+    cd web
+    rm -rf themes
+    git clone --depth 1 --recursive https://github.com/gohugoio/hugoThemes.git themes
+    cd ..
+fi
+
 # link all themes to web themes
-cd web
-rm -rf themes
-git clone --depth 1 --recursive https://github.com/gohugoio/hugoThemes.git themes
-cd ../website
-exit
 cd website
+
 ## declare an array variable
 declare -a arr=(
     "Huny-B.com"
@@ -39,28 +45,30 @@ for i in "${arr[@]}"
 do
     cd $i
         echo $i
-#        rm -rf themes
-       rm -rf public
-#        git add -A
-#        git commit -m $i
+        if [[ ! -d "themes/.git" ]]; then 
+            rm -rf themes
+            git add -A
+            git commit -m $i
 		
-#        if windows; then
-#           cmd <<< "mklink /D themes \..\..\web\themes" > /dev/null
+            if windows; then
+               cmd <<< "mklink /D themes \..\..\web\themes" > /dev/null
+            else
+               ln -s ../../web/themes
+            fi
+		fi
+        
+#		if windows; then
+#		   rm -rf public
+#           git add -A
+#		   git commit -m pub$i
+#           mkdir public
+#		   git add -A
+#		   git commit -m pub$i
+#		   rm -rf public
+#		   git submodule add -b master git@github.com:Huny-B-CBD-Oil/$i.git public 
 #        else
-#           ln -s ../../web/themes
+#         git submodule add -b master git@github.com:Huny-B-CBD-Oil/$i.git public 
 #        fi
-		
-		if windows; then
-		   git add -A
-		   git commit -m pub$i
-           mkdir public
-		   git add -A
-		   git commit -m pub$i
-		   rm -rf public
-		   git submodule add -b master git@github.com:Huny-B-CBD-Oil/$i.git public 
-        else
-         git submodule add -b master git@github.com:Huny-B-CBD-Oil/$i.git public 
-        fi
 		
     cd ..
 done
